git: create a custom git commit template
 __ISSUE:__
Using "git commit -m "fixed $PATH bugs" sells a poor mindset, and
makes the commit feel more like a one-off argument than a history
page.

 __DESCRIPTION:__
To produce a more informative git history, we should
be using a .gitmessage template file. We can think of this feature of
Git as a tool to encourage us to develop a better habit of writing
useful commit messages.

 ___STEPS TO TEST:__
Git provides a way to create a commit message template via the
`.gitmessage` file. Setting up a custom commit template is easy.
First, we will need to edit our ".gitconfig" file to point git to the template.
We can create the template in the file .gitmessage in the root folder,
but you can use every file name in any location you like. If the
specified path-to-template-file is not absolute, it will be treated
relative to the repository root. There are two options to edit
.gitconfig, via the command line or manually in your editor.

I used the command-line approach:
  git config --global commit.template ~/.gitmessage

 __REFERENCES:__ #
  * Git Gud: Create a .gitmessage
  https://dev.to/vetswhocode/git-gud-create-a-gitmessage-4ibj

  * Create A Custom Git Commit Template
  https://alex-wasik.medium.com/create-a-custom-git-commit-template-84468232a459

  * 5 Useful Tips For A Better Commit Message
  https://thoughtbot.com/blog/5-useful-tips-for-a-better-commit-message